add, usercontroller registers to db
