+++ controllers/UserController.ts
@@ -1,131 +1,43 @@
 import express, { Request, Response, Router, NextFunction } from "express";
-
-import gravatar from "gravatar";
 import bcrypt from "bcryptjs";
 import jwt from "jsonwebtoken";
+//custom imports
+import User from "../models/user";
 import asyncHandler from "../middlewares/async";
 import ErrorResponse from "../utils/ErrorResponse";
-import { pool } from "../config/db";
-
+//types
 import { ec } from "elliptic";
-
 const ecGenerate = new ec("secp256k1");
 
 class UserController {
-  getUsers = asyncHandler(async (req: any, res: any, next: any) => {
-    if (
-      req?.headers?.authorization &&
-      req?.headers?.authorization?.includes("Bearer")
-    ) {
-      const token = req.headers.authorization.slice(
-        req.headers.authorization.indexOf("Bearer") + 7
-      );
-
-      const { rows } = await pool.query(
-        `SELECT * FROM accounts WHERE token = $1`,
-        [token]
-      );
-
-      const user = rows[0] || false;
-
-      if (user) {
-        const users = await pool.query("SELECT * FROM accounts");
-
-        res.json(users.rows);
-      } else {
-        const users = await pool.query("SELECT user_id FROM accounts");
-
-        res.json(users.rows);
-      }
-    } else {
-      const users = await pool.query("SELECT user_id FROM accounts");
-
-      res.json(users.rows);
+  /**
+   * Create new user method, returns object with success message
+   * authentication status true on user creation
+   * the token generated
+   *
+   * @param req
+   * @param res
+   * @param next
+   * @returns object
+   */
+  register = asyncHandler(async (req: any, res: any, next: any) => {
+    //check if password matches the confirmation password 2 form inputs
+    if (req.body.password !== req.body.confPassword) {
+      return res.status(400).send({
+        message: "password provided does not match the confirmation password",
+      });
+    }
+    try {
+      await User.create(req.body);
+      res.status(201).json({
+        message: "User Created",
+        auth: true,
+        token: "token",
+      });
+    } catch (error) {
+      res.status(400).send(error);
     }
   });
-
-  register = asyncHandler(
-    async (req: Request, res: Response, next: NextFunction) => {
-      const { name, email, password, accountType } = req.body;
-
-      console.log(req.body);
-
-      const { rows } = await pool.query(
-        `SELECT * FROM accounts WHERE email = $1`,
-        [email]
-      );
-
-      let user = (await rows[0]) || false;
-
-      if (user) {
-        return next(new ErrorResponse("User already exists.", 422));
-      } else if (!email || !email.includes("@")) {
-        return next(new ErrorResponse("Enter @ address.", 422));
-      }
-
-      let avatar = gravatar.url(email, {
-        s: "200",
-        r: "pg",
-        d: "mm",
-      });
-
-      if (avatar && avatar.toString() && !avatar.toString().includes("https")) {
-        avatar = "https:" + avatar.toString();
-      }
-
-      const userName = (await name) || email.slice(0, email.indexOf("@"));
-
-      // encrypt password using bcrypt
-      const salt = await bcrypt.genSalt(10);
-
-      const safePassword = await bcrypt.hash(password, salt);
-
-      const key = ecGenerate.genKeyPair();
-      const publicKey = key.getPublic("hex");
-      const privateKey = key.getPrivate("hex");
-
-      const users = await pool.query(
-        `INSERT INTO accounts (name, email, password, avatar, public_key, private_key, account_type) VALUES($1, $2, $3, $4, $5, $6, $7) RETURNING *`,
-        [
-          userName,
-          email,
-          safePassword,
-          avatar || "",
-          publicKey,
-          privateKey,
-          accountType,
-        ]
-      );
-
-      user = await users?.rows[0];
-
-      const payload = {
-        user: {
-          id: user.user_id,
-        },
-      };
-
-      const JWTSecretKey: any = process.env["jwtSecret"];
-
-      return jwt.sign(
-        payload,
-        JWTSecretKey,
-        { expiresIn: 360000 },
-        async (err, token) => {
-          if (err) {
-            return next(new ErrorResponse(err.message, 422));
-          }
-
-          await pool.query(`UPDATE accounts SET token = $1 WHERE email = $2`, [
-            token,
-            email,
-          ]);
-
-          res.json({ success: true, token });
-        }
-      );
-    }
-  );
 }
 
 export default UserController;
