+++ routes/api/users.ts
@@ -1,25 +1,19 @@
-
-
-import express, {Request, Response, Router, NextFunction} from 'express';
-
+import express, { Request, Response, Router, NextFunction } from "express";
 
 const router: Router = express.Router();
 
+import UserController from "../../controllers/UserController";
 
-import UserController from '../../controllers/UserController';
-
-const userController = new UserController;
-
+const userController = new UserController();
 
 //route get    api/users
 //description  get users
 //access       private
-router.get('/', userController.getUsers);
+// router.get("/", userController.getUsers);
 
 //route get    api/users
 //description  register
 //access       private
-router.post('/', userController.register);
-
+router.post("/", userController.register); // should be public
 
 export default router;
